removed new_app_manager and return to app_manager
